requirements.txt
autogen-agentchat
autogen-ext[azure]
azure-search-documents==11.6.0b8
azure-identity
tiktoken
openai
python-dotenv
black
ipykernel
jupyter
chainlit